refactor(OBS-1): check:architecture exit 0 — brain-dreaming split + observer-memory port
brain-dreaming.ts 749->622 via brain-dream-proposer.ts leaf extraction
(exports preserved, no cycle); observer-active-memory.ts consumes an
ObserverActiveMemoryReadPort instead of importing postgres schema+drizzle
directly (memory-layer boundary restored per the app-query pattern). Zero
behavior change.
